requirements.txt
# Core dependencies for Metisara
jira>=3.0.0
python-dotenv>=0.19.0

# Standard library packages (no installation needed)
# configparser - included in Python 3+
# pathlib - included in Python 3.4+
# argparse - included in Python 2.7+
# subprocess - included in Python 2.4+
# csv - included in Python 2.3+
# json - included in Python 2.6+
# zipfile - included in Python 1.6+
# shutil - included in Python 1.4+
# datetime - included in Python 2.3+